+++ src/main/java/io/supertokens/storage/postgresql/Start.java
@@ -644,15 +644,6 @@ public class Start
         }
     }
 
-    private String[] getAllNonExpiredSessionHandlesForUser(AppIdentifier appIdentifier, String userId)
-            throws StorageQueryException {
-        try {
-            return SessionQueries.getAllNonExpiredSessionHandlesForUser(this, appIdentifier, userId);
-        } catch (SQLException e) {
-            throw new StorageQueryException(e);
-        }
-    }
-
     @Override
     public void deleteAllExpiredSessions() throws StorageQueryException {
         try {
@@ -859,11 +850,17 @@ public class Start
             throws StorageQueryException {
         // check if the input userId is being used in nonAuthRecipes.
         if (className.equals(SessionStorage.class.getName())) {
-            String[] sessionHandlesForUser = getAllNonExpiredSessionHandlesForUser(appIdentifier, userId);
-            return sessionHandlesForUser.length > 0;
+            try {
+                return SessionQueries.doesNonExpiredSessionExistForUser(this, appIdentifier, userId);
+            } catch (SQLException e) {
+                throw new StorageQueryException(e);
+            }
         } else if (className.equals(UserRolesStorage.class.getName())) {
-            String[] roles = getRolesForUser(appIdentifier, userId);
-            return roles.length > 0;
+            try {
+                return UserRolesQueries.doesUserHaveAnyRole(this, appIdentifier, userId);
+            } catch (SQLException e) {
+                throw new StorageQueryException(e);
+            }
         } else if (className.equals(UserMetadataStorage.class.getName())) {
             JsonObject userMetadata = getUserMetadata(appIdentifier, userId);
             return userMetadata != null;
@@ -875,8 +872,7 @@ public class Start
             }
         } else if (className.equals(TOTPStorage.class.getName())) {
             try {
-                TOTPDevice[] devices = TOTPQueries.getDevices(this, appIdentifier, userId);
-                return devices.length > 0;
+                return TOTPQueries.doesUserHaveAnyDevice(this, appIdentifier, userId);
             } catch (SQLException e) {
                 throw new StorageQueryException(e);
             }
@@ -2720,15 +2716,6 @@ public class Start
             StorageQueryException {
         try {
             return UserRolesQueries.getRolesForUser(this, tenantIdentifier, userId);
-        } catch (SQLException e) {
-            throw new StorageQueryException(e);
-        }
-    }
-
-    private String[] getRolesForUser(AppIdentifier appIdentifier, String userId) throws
-            StorageQueryException {
-        try {
-            return UserRolesQueries.getRolesForUser(this, appIdentifier, userId);
         } catch (SQLException e) {
             throw new StorageQueryException(e);
         }